apply clippy suggestions
