+++ src/db/car/any.rs
@@ -5,23 +5,23 @@ use fvm_ipld_blockstore::Blockstore;
 use std::io::{Cursor, Error, ErrorKind, Read, Result, Seek};
 
 pub enum AnyCar<ReaderT> {
-    PlainCar(super::PlainCar<ReaderT>),
-    ForestCar(super::ForestCar<ReaderT>),
-    MemoryCar(super::PlainCar<Cursor<Vec<u8>>>),
+    Plain(super::PlainCar<ReaderT>),
+    Forest(super::ForestCar<ReaderT>),
+    Memory(super::PlainCar<Cursor<Vec<u8>>>),
 }
 
 impl<ReaderT: Read + Seek> AnyCar<ReaderT> {
     pub fn new(mk_reader: impl Fn() -> ReaderT + Clone + 'static) -> Result<Self> {
         if let Ok(forest_car) = super::ForestCar::new(mk_reader.clone()) {
-            return Ok(AnyCar::ForestCar(forest_car));
+            return Ok(AnyCar::Forest(forest_car));
         }
         if let Ok(plain_car) = super::PlainCar::new(mk_reader()) {
-            return Ok(AnyCar::PlainCar(plain_car));
+            return Ok(AnyCar::Plain(plain_car));
         }
         if let Ok(decompressed) = zstd::stream::decode_all(mk_reader()) {
             let mem_reader = Cursor::new(decompressed);
             if let Ok(mem_car) = super::PlainCar::new(mem_reader) {
-                return Ok(AnyCar::MemoryCar(mem_car));
+                return Ok(AnyCar::Memory(mem_car));
             }
         }
         Err(Error::new(
@@ -32,9 +32,9 @@ impl<ReaderT: Read + Seek> AnyCar<ReaderT> {
 
     pub fn roots(&self) -> Vec<Cid> {
         match self {
-            AnyCar::ForestCar(forest) => forest.roots(),
-            AnyCar::PlainCar(plain) => plain.roots(),
-            AnyCar::MemoryCar(mem) => mem.roots(),
+            AnyCar::Forest(forest) => forest.roots(),
+            AnyCar::Plain(plain) => plain.roots(),
+            AnyCar::Memory(mem) => mem.roots(),
         }
     }
 }
@@ -45,17 +45,17 @@ where
 {
     fn get(&self, k: &Cid) -> anyhow::Result<Option<Vec<u8>>> {
         match self {
-            AnyCar::ForestCar(forest) => forest.get(k),
-            AnyCar::PlainCar(plain) => plain.get(k),
-            AnyCar::MemoryCar(mem) => mem.get(k),
+            AnyCar::Forest(forest) => forest.get(k),
+            AnyCar::Plain(plain) => plain.get(k),
+            AnyCar::Memory(mem) => mem.get(k),
         }
     }
 
     fn put_keyed(&self, k: &Cid, block: &[u8]) -> anyhow::Result<()> {
         match self {
-            AnyCar::ForestCar(forest) => forest.put_keyed(k, block),
-            AnyCar::PlainCar(plain) => plain.put_keyed(k, block),
-            AnyCar::MemoryCar(mem) => mem.put_keyed(k, block),
+            AnyCar::Forest(forest) => forest.put_keyed(k, block),
+            AnyCar::Plain(plain) => plain.put_keyed(k, block),
+            AnyCar::Memory(mem) => mem.put_keyed(k, block),
         }
     }
 }
