+++ src/utils/db/car_index/tests.rs
@@ -1,34 +1,9 @@
 // Copyright 2019-2023 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 use super::*;
+use ahash::{AHashMap, AHashSet};
 use quickcheck_macros::quickcheck;
-use std::collections::{HashMap, HashSet};
 use std::io::{Cursor, Read, Seek};
-
-// #[test]
-// fn show_misses_and_collisions() {
-//     let table = ProbingHashtableBuilder::new(
-//         &(1..=100)
-//             .map(|i| (Cid::from_cbor_blake2b256(&i).unwrap(), i))
-//             .collect::<Vec<_>>(),
-//     );
-//     dbg!(table.read_misses());
-//     dbg!(table.collisions);
-// }
-
-// #[test]
-// fn show_layout() {
-//     let table = ProbingHashtableBuilder::new_raw(&[
-//         (Hash(0), Position::decode(0)),
-//         (Hash(1), Position::decode(0)),
-//         (Hash(2), Position::decode(0)),
-//         (Hash(3), Position::decode(0)),
-//         (Hash(6), Position::decode(0)),
-//         (Hash(6), Position::decode(0)),
-//         (Hash(6), Position::decode(0)),
-//     ]);
-//     dbg!(table);
-// }
 
 fn query(table: &mut CarIndex<impl Read + Seek>, key: Hash) -> Vec<BlockPosition> {
     table.lookup_hash(key).unwrap().into_vec()
@@ -41,14 +16,14 @@ fn mk_table(entries: &[(Hash, BlockPosition)]) -> CarIndex<Cursor<Vec<u8>>> {
     CarIndex::open(Cursor::new(store), 0).unwrap()
 }
 
-fn mk_map(entries: &[(Hash, BlockPosition)]) -> HashMap<Hash, HashSet<BlockPosition>> {
-    let mut map = HashMap::with_capacity(entries.len());
+fn mk_map(entries: &[(Hash, BlockPosition)]) -> AHashMap<Hash, AHashSet<BlockPosition>> {
+    let mut map = AHashMap::with_capacity(entries.len());
     for (hash, position) in entries.iter().copied() {
         map.entry(hash)
-            .and_modify(|set: &mut HashSet<BlockPosition>| {
+            .and_modify(|set: &mut AHashSet<BlockPosition>| {
                 set.insert(position);
             })
-            .or_insert(HashSet::from([position]));
+            .or_insert(AHashSet::from([position]));
     }
     map
 }
@@ -66,7 +41,7 @@ fn lookup_wide(entries: Vec<(Hash, BlockPosition)>) {
     let map = mk_map(&entries);
     let mut table = mk_table(&entries);
     for (&hash, value_set) in map.iter() {
-        assert_eq!(&HashSet::from_iter(query(&mut table, hash)), value_set);
+        assert_eq!(&AHashSet::from_iter(query(&mut table, hash)), value_set);
     }
 }
 
@@ -79,7 +54,7 @@ fn lookup_narrow(mut entries: Vec<(Hash, BlockPosition)>) {
     let map = mk_map(&entries);
     let mut table = mk_table(&entries);
     for (&hash, value_set) in map.iter() {
-        assert_eq!(&HashSet::from_iter(query(&mut table, hash)), value_set);
+        assert_eq!(&AHashSet::from_iter(query(&mut table, hash)), value_set);
     }
 }
 
@@ -95,7 +70,7 @@ fn lookup_clash_all(mut entries: Vec<(Hash, BlockPosition)>) {
     let map = mk_map(&entries);
     let mut table = mk_table(&entries);
     for (&hash, value_set) in map.iter() {
-        assert_eq!(&HashSet::from_iter(query(&mut table, hash)), value_set);
+        assert_eq!(&AHashSet::from_iter(query(&mut table, hash)), value_set);
     }
 }
 
@@ -112,6 +87,6 @@ fn lookup_clash_many(mut entries: Vec<(Hash, BlockPosition)>) {
     let map = mk_map(&entries);
     let mut table = mk_table(&entries);
     for (&hash, value_set) in map.iter() {
-        assert_eq!(&HashSet::from_iter(query(&mut table, hash)), value_set);
+        assert_eq!(&AHashSet::from_iter(query(&mut table, hash)), value_set);
     }
 }
